+++ libs/zhi-gitlab-middleware/src/lib/commonGitlabClient.ts
@@ -108,9 +108,10 @@ class CommonGitlabClient {
    *
    * @param filePath - 文件路径
    * @param content - 文件内容
+   * @param encoding - base64 或者 text
    * @returns 创建文件的响应
    */
-  public async createRepositoryFile(filePath: string, content: string): Promise<any> {
+  public async createRepositoryFile(filePath: string, content: string, encoding = "text"): Promise<any> {
     const id = `${this.user}/${this.repo}`
     const endpointUrl =
       `/api/v4/projects/${encodeURIComponent(id)}/repository/files/` + `${encodeURIComponent(filePath)}`
@@ -118,6 +119,7 @@ class CommonGitlabClient {
       branch: this.branch,
       author_email: this.authorEmail,
       author_name: this.authorName,
+      encoding: encoding,
       content: content,
       commit_message: this.commitMessage,
     }